+++ CHANGES.md
@@ -32,6 +32,7 @@ Changes:
 
 - Retrieve lower court information in `delaware` scrapers #1569
 - Retrieve lower court information in `idaho` #1569
+- Retrieve lower court information in `ky` #1569
 - Retrieve lower court information in `la` #1569
 
 Fixes: